logging: handler levels and performance optimization
The logging level is filtered by the Loggers, there is no need to restrict it at the handler
level. It is also useful to let the handler level on ALL to allow specifying finer levels for
specific classes using a logging file configuration.

Some optimizations are made to avoid computing String.format and the stack trace.